rgw: dmclock: move client config and counters to its own header
