+++ src/rgw/rgw_dmclock_scheduler.h
@@ -3,8 +3,8 @@
 /*
  * Ceph - scalable distributed file system
  *
- * Copyright (C) 2018 Red Hat, Inc.
- *
+ * Copyright (C) 2019 Red Hat, Inc.
+ *           (C) 2019 SUSE LLC
  * This is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License version 2.1, as published by the Free Software
@@ -18,34 +18,10 @@
 #include "common/ceph_time.h"
 #include "common/ceph_context.h"
 #include "common/config.h"
-#include "common/perf_counters.h"
-#include "rgw_dmclock.h"
 #include "common/async/yield_context.h"
+#include "rgw_dmclock.h"
 
 namespace rgw::dmclock {
-
-namespace queue_counters {
-
-enum {
-  l_first = 427150,
-  l_qlen,
-  l_cost,
-  l_res,
-  l_res_cost,
-  l_prio,
-  l_prio_cost,
-  l_limit,
-  l_limit_cost,
-  l_cancel,
-  l_cancel_cost,
-  l_res_latency,
-  l_prio_latency,
-  l_last,
-};
-
-PerfCountersRef build(CephContext *cct, const std::string& name);
-
-} // namespace queue_counters
 
 /// function to provide client counters
 using GetClientCounters = std::function<PerfCounters*(client_id)>;
@@ -84,6 +60,23 @@ private:
 
 using SchedulerCompleter = Completer<std::function<void()>>;
 
+enum class scheduler_t {
+  none,
+  throttler,
+  dmclock
+};
+
+inline scheduler_t get_scheduler_t(CephContext *cct)
+{
+  const auto scheduler_type = cct->_conf.get_val<std::string>("rgw_scheduler_type");
+  if (scheduler_type == "dmclock")
+    return scheduler_t::dmclock;
+  else if (scheduler_type == "throttler")
+    return scheduler_t::throttler;
+  else
+    return scheduler_t::none;
+}
+
 class Scheduler  {
 public:
   auto schedule_request(const client_id& client, const ReqParams& params,
@@ -102,47 +95,6 @@ private:
 				    optional_yield) = 0;
 };
 
-/// array of per-client counters to serve as GetClientCounters
-class ClientCounters {
-  std::array<PerfCountersRef, static_cast<size_t>(client_id::count)> clients;
- public:
-  ClientCounters(CephContext *cct);
-
-  PerfCounters* operator()(client_id client) const {
-    return clients[static_cast<size_t>(client)].get();
-  }
-};
-
-struct ClientSum {
-  uint64_t count{0};
-  Cost cost{0};
-};
-
-constexpr auto client_count = static_cast<size_t>(client_id::count);
-using ClientSums = std::array<ClientSum, client_count>;
-
-void inc(ClientSums& sums, client_id client, Cost cost);
-void on_cancel(PerfCounters *c, const ClientSum& sum);
-void on_process(PerfCounters* c, const ClientSum& rsum, const ClientSum& psum);
-
-/// a simple wrapper to hold client config. objects needed to construct a
-/// scheduler instance, the primary utility of this being to optionally
-/// construct scheduler only when configured in the frontends.
-class optional_scheduler_ctx {
-  std::optional<ClientConfig> clients;
-  std::optional<ClientCounters> counters;
-public:
-  optional_scheduler_ctx(CephContext *cct) {
-    if(cct->_conf.get_val<bool>("rgw_dmclock_enabled")){
-      clients.emplace(ClientConfig(cct));
-      counters.emplace(ClientCounters(cct));
-    }
-  }
-  operator bool() const noexcept { return counters && clients; }
-  // both the get functions below will throw
-  ClientCounters& get_counters() { return counters.value(); }
-  ClientConfig& get_clients() { return clients.value(); }
-};
 } // namespace rgw::dmclock
 
 #endif // RGW_DMCLOCK_SCHEDULER_H
