+++ src/rgw/rgw_op.h
@@ -117,6 +117,8 @@ public:
 
 void rgw_bucket_object_pre_exec(struct req_state *s);
 
+namespace dmc = rgw::dmclock;
+
 /**
  * Provide the base class for all ops.
  */
@@ -201,9 +203,8 @@ public:
   CephContext* get_cct() const override { return s->cct; }
   unsigned get_subsys() const override { return ceph_subsys_rgw; }
 
-  using client_id = rgw::dmclock::client_id;
-  virtual client_id dmclock_client() { return client_id::metadata; }
-  virtual rgw::dmclock::Cost dmclock_cost() { return 1; }
+  virtual dmc::client_id dmclock_client() { return dmc::client_id::metadata; }
+  virtual dmc::Cost dmclock_cost() { return 1; }
 };
 
 class RGWDefaultResponseOp : public RGWOp {
@@ -360,7 +361,7 @@ public:
     *filter = nullptr;
     return 0;
   }
-  client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 class RGWGetObj_CB : public RGWGetObj_Filter
@@ -489,7 +490,7 @@ public:
   const char* name() const override { return "bulk_delete"; }
   RGWOpType get_type() override { return RGW_OP_BULK_DELETE; }
   uint32_t op_mask() override { return RGW_OP_TYPE_DELETE; }
-  rgw::dmclock::client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 inline ostream& operator<<(ostream& out, const RGWBulkDelete::acct_path_t &o) {
@@ -566,7 +567,7 @@ public:
   uint32_t op_mask() override {
     return RGW_OP_TYPE_WRITE;
   }
-  client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 }; /* RGWBulkUploadOp */
 
 
@@ -1106,7 +1107,7 @@ public:
   const char* name() const override { return "put_obj"; }
   RGWOpType get_type() override { return RGW_OP_PUT_OBJ; }
   uint32_t op_mask() override { return RGW_OP_TYPE_WRITE; }
-  client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 class RGWPostObj : public RGWOp {
@@ -1160,7 +1161,7 @@ public:
   const char* name() const override { return "post_obj"; }
   RGWOpType get_type() override { return RGW_OP_POST_OBJ; }
   uint32_t op_mask() override { return RGW_OP_TYPE_WRITE; }
-  client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 class RGWPutMetadataAccount : public RGWOp {
@@ -1292,7 +1293,7 @@ public:
   RGWOpType get_type() override { return RGW_OP_DELETE_OBJ; }
   uint32_t op_mask() override { return RGW_OP_TYPE_DELETE; }
   virtual bool need_object_expiration() { return false; }
-  rgw::dmclock::client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 class RGWCopyObj : public RGWOp {
@@ -1384,7 +1385,7 @@ public:
   const char* name() const override { return "copy_obj"; }
   RGWOpType get_type() override { return RGW_OP_COPY_OBJ; }
   uint32_t op_mask() override { return RGW_OP_TYPE_WRITE; }
-  rgw::dmclock::client_id dmclock_client() override { return client_id::data; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::data; }
 };
 
 class RGWGetACLs : public RGWOp {
@@ -2186,7 +2187,7 @@ public:
   virtual int get_params() = 0;
   void execute() override;
   const char* name() const override { return "get_cluster_stat"; }
-  client_id dmclock_client() override { return client_id::admin; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::admin; }
 };
 
 static inline int parse_value_and_bound(
