+++ src/rgw/rgw_rest.h
@@ -488,7 +488,7 @@ public:
   virtual int check_caps(RGWUserCaps& caps)
     { return -EPERM; } /* should to be implemented! */
   int verify_permission() override;
-  client_id dmclock_client() override { return client_id::admin; }
+  dmc::client_id dmclock_client() override { return dmc::client_id::admin; }
 };
 
 class RGWHandler_REST : public RGWHandler {
